refactor(dashboard-user): move user info to user module
