+++ apps/dashboard/src/modules/user/components/balance/BalanceWithTopup.vue
@@ -59,7 +59,7 @@
 
 <script setup lang="ts">
 import CardComponent from '@/components/CardComponent.vue';
-import TopupModal from '@/modules/home/components/TopupModal.vue';
+import BalanceTopupModal from '@/modules/user/components/balance/BalanceTopupModal.vue';
 import { useAuthStore, useUserStore } from '@sudosos/sudosos-frontend-common';
 import { computed, ref, onMounted, type Ref, watch } from 'vue';
 import type { BalanceResponse } from '@sudosos/sudosos-client';
